Adding stats command, can now handle responses
We have basic functionality on handling events that we are sent from the
game client now, so we can implement commands such as `!stats`. There
are still some issues, such as none of the 'join_game' events that we
expect are coming in to the scoket? The leave game events are though...